@@ -1,7 +1,7 @@
 package bankaccount;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
 import java.time.LocalDate;
@@ -11,7 +11,7 @@ public class BankAccountFeature {
     private Printer printer;
     private DateService dateService;
 
-    @Before
+    @BeforeEach
     public void setUp() {
         printer = Mockito.mock(Printer.class);
         dateService = Mockito.mock(DateService.class);
